+++ src/main/array.c
@@ -1189,7 +1189,7 @@ SEXP attribute_hidden do_aperm(SEXP call, SEXP op, SEXP args, SEXP rho)
 
     /* check the permutation */
 
-    int *pp = (int *)R_alloc(n, sizeof(int));
+    int *pp = (int *)R_alloc((size_t)n, sizeof(int));
     perm = CADR(args);
     if (length(perm) == 0)
     {
@@ -1230,7 +1230,7 @@ SEXP attribute_hidden do_aperm(SEXP call, SEXP op, SEXP args, SEXP rho)
             error(_("'perm' is of wrong length"));
     }
 
-    int *iip = (int *)R_alloc(n, sizeof(int));
+    int *iip = (int *)R_alloc((size_t)n, sizeof(int));
     for (i = 0; i < n; iip[i++] = 0)
         ;
     for (i = 0; i < n; i++)
@@ -1244,7 +1244,7 @@ SEXP attribute_hidden do_aperm(SEXP call, SEXP op, SEXP args, SEXP rho)
 
     /* create the stride object and permute */
 
-    int *stride = (int *)R_alloc(n, sizeof(int));
+    int *stride = (int *)R_alloc((size_t)n, sizeof(int));
     for (iip[0] = 1, i = 1; i < n; i++)
         iip[i] = iip[i - 1] * isa[i - 1];
     for (i = 0; i < n; i++)
